rename battle.py to cli.py
